make: got rid of strict aliasing warnings
Unfortunately I still don't know a good/performant way to get rid of the
strict aliasing warnings on newer GCC's. Creating a union is not really
desirable as it's been reported to lower performance and creates a lot more
typing work. The cases that are in the codebase now are false positives
though.