drawlist: first draft
Trying to find a good format for an easily-sortable drawlist, as proposed in
many articles online. It appears that bitfields are fraught with peril.
Endianness, compiler, ISA all play a role in making it a vaguely defined
cesspool. Perhaps I better shift to raw bit operations and leave the syntax
sugar of bitfields behing...