drawlist: reorder bitfields, model less important
Because we recognize the correlation between model and parameters, if model
is the same then parameters will also likely be. So we "gain" more batching
by putting the model bitfield behind the others. This also gives me an idea
to run a statistical analysis on generated drawlist keys for large amounts
of real-world models, and deduce based on the resulting correlations and
some intuition or even benchmarks the best bitfield ordering. 't would be
cool to include that as a script that can be ran for a target system, which
would generate an adjusted drawlist.h file. Ah... dreams.